Change UServer::lastTime_ type to utime_t.
	* include/kernel/userver.hh,
	* include/kernel/userver.hxx: Here.
	* src/uconnection.cc: Remove useless cast on lastTime().

+++ src/uconnection.cc
@@ -424,7 +424,7 @@ UConnection::make_prefix (const char* tag) const
 {
   std::ostringstream o;
   char fill = o.fill('0');
-  o << '[' << std::setw(8) << (int) server_->lastTime();
+  o << '[' << std::setw(8) << server_->lastTime();
   o.fill(fill);
   if (tag && strlen(tag))
     o << ':' << tag;
